+++ dolfyn/adv/turbulence.py
@@ -103,7 +103,7 @@ class ADVBinner(VelBinner):
 
         Parameters
         ----------
-        veldat   : xarray.DataArray
+        veldat : xarray.DataArray
           The raw 3D velocity data.
         freq_units : string
           Frequency units of the returned spectra in either Hz or rad/s 
@@ -125,25 +125,29 @@ class ADVBinner(VelBinner):
           different cross-spectra: 'uv', 'uw', 'vw'.
         """
 
-        fs = self._parse_fs(fs)
+        fs_in = self._parse_fs(fs)
         n_fft = self._parse_nfft_coh(n_fft_coh)
         time = self.mean(veldat.time.values)
         veldat = veldat.values
+        if len(np.shape(veldat)) != 2:
+            raise Exception("This function is only valid for calculating TKE using "
+                            "the 3D velocity vector from an ADV.")
 
-        out = np.empty(self._outshape_fft(veldat[:3].shape, n_fft=n_fft),
+        out = np.empty(self._outshape_fft(veldat[:3].shape, n_fft=n_fft, n_bin=n_bin),
                        dtype='complex')
 
         # Create frequency vector, also checks whether using f or omega
         if 'rad' in freq_units:
-            fs = 2*np.pi*fs
+            fs = 2*np.pi*fs_in
             freq_units = 'rad s-1'
             units = 'm2 s-1 rad-1'
         else:
+            fs = fs_in
             freq_units = 'Hz'
             units = 'm2 s-2 Hz-1'
-        coh_freq = xr.DataArray(self.calc_freq(units=freq_units, coh=True),
-                                dims=['freq'],
-                                name='freq',
+        coh_freq = xr.DataArray(self.calc_freq(fs=fs_in, units=freq_units, n_fft=n_fft, coh=True),
+                                dims=['coh_freq'],
+                                name='coh_freq',
                                 attrs={'units': freq_units,
                                       'long_name': 'FFT Frequency Vector',
                                       'coverage_content_type': 'coordinate'}
@@ -152,29 +156,154 @@ class ADVBinner(VelBinner):
         for ip, ipair in enumerate(self._cross_pairs):
             out[ip] = self.calc_csd_base(veldat[ipair[0]],
                                          veldat[ipair[1]],
-                                         n_bin=n_bin,
-                                         n_fft=n_fft,
-                                         window=window)
+                                     fs=fs,
+                                     window=window,
+                                     n_bin=n_bin,
+                                     n_fft=n_fft)
 
         csd = xr.DataArray(out.astype('complex64'),
                            coords={'C': self.C,
                                    'time': time,
-                                   'freq': coh_freq},
-                           dims=['C', 'time', 'freq'],
+                                   'coh_freq': coh_freq},
+                           dims=['C', 'time', 'coh_freq'],
                            attrs={'units': units, 
                                   'n_fft_coh': n_fft,
                                   'long_name': 'Cross Spectral Density',
                                   'standard_name': 'cross_spectral_density_of_sea_water_velocity'})
-        csd['freq'].attrs['units'] = freq_units
+        csd['coh_freq'].attrs['units'] = freq_units
 
         return csd
+
+    def calc_doppler_noise(self, psd, pct_fN=0.8):
+        """Calculate bias due to Doppler noise using the noise floor
+        of the velocity spectra.
+
+        Parameters
+        ----------
+        psd : xarray.DataArray (dir, time, f)
+          The ADV velocity spectra
+        pct_fN : float
+          Percent of Nyquist frequency to calculate characeristic frequency
+
+        Returns
+        -------
+        doppler_noise (xarray.DataArray): 
+          Doppler noise level in units of m/s
+
+        Notes
+        -----
+        Approximates bias from
+
+        .. :math: \\sigma^{2}_{noise} = N x f_{c}
+
+        where :math: `\\sigma_{noise}` is the bias due to Doppler noise,
+        `N` is the constant variance or spectral density, and `f_{c}`
+        is the characteristic frequency.
+
+        The characteristic frequency is then found as 
+
+        .. :math: f_{c} = pct_fN * (f_{s}/2)
+
+        where `f_{s}/2` is the Nyquist frequency.
+
+
+        Richard, Jean-Baptiste, et al. "Method for identification of Doppler noise 
+        levels in turbulent flow measurements dedicated to tidal energy." International 
+        Journal of Marine Energy 3 (2013): 52-64.
+
+        Thiébaut, Maxime, et al. "Investigating the flow dynamics and turbulence at a 
+        tidal-stream energy site in a highly energetic estuary." Renewable Energy 195 
+        (2022): 252-262.
+        """
+        
+        # Characteristic frequency set to 80% of Nyquist frequency
+        fN = self.fs/2
+        fc = pct_fN * fN
+
+        # Get units right
+        if psd.freq.units == "Hz":
+            f_range = slice(fc, fN)
+        else:
+            f_range = slice(2*np.pi*fc, 2*np.pi*fN)
+
+        # Noise floor
+        N2 = psd.sel(freq=f_range) * psd.freq.sel(freq=f_range)
+        noise_level = np.sqrt(N2.mean(dim='freq'))
+
+        return xr.DataArray(
+            noise_level.values.astype('float32'),
+            dims=['dir', 'time'],
+            attrs={'units': 'm/s',
+                   'long_name': 'Doppler Noise Level',
+                   'description': 'Doppler noise level calculated '
+                   'from PSD white noise'})
+
+    def check_turbulence_cascade_slope(self, psd, freq_range=[6.28, 12.57]):
+        """This function calculates the slope of the PSD, the power spectra 
+        of velocity, within the given frequency range. The purpose of this
+        function is to check that the region of the PSD containing the 
+        isotropic turbulence cascade decreases at a rate of :math:`f^{-5/3}`.
+
+        Parameters
+        ----------
+        psd : xarray.DataArray ([time,] freq)
+          The power spectral density (1D or 2D)
+        freq_range : iterable(2) (default: [6.28, 12.57])
+          The range over which the isotropic turbulence cascade occurs, in 
+          units of the psd frequency vector (Hz or rad/s)
+
+        Returns
+        -------
+        (m, b): tuple (slope, y-intercept)
+          A tuple containing the coefficients of the log-adjusted linear 
+          regression between PSD and frequency 
+
+        Notes
+        -----
+        Calculates slope based on the `standard` formula for dissipation:
+
+        .. math:: S(k) = \\alpha \\epsilon^{2/3} k^{-5/3} + N
+
+        The slope of the isotropic turbulence cascade, which should be 
+        equal to :math:`k^{-5/3}` or :math:`f^{-5/3}`, where k and f are 
+        the wavenumber and frequency vectors, is estimated using linear 
+        regression with a log transformation:
+
+        .. math:: log10(y) = m*log10(x) + b
+
+        Which is equivalent to
+
+        .. math:: y = 10^{b} x^{m}
+        
+        Where :math:`y` is S(k) or S(f), :math:`x` is k or f, :math:`m` 
+        is the slope (ideally -5/3), and :math:`10^{b}` is the intercept of 
+        y at x^m=1.
+        """
+
+        idx = np.where((freq_range[0] < psd.freq) & (psd.freq < freq_range[1]))
+        idx = idx[0]
+
+        x = np.log10(psd['freq'].isel(freq=idx))
+        y = np.log10(psd.isel(freq=idx))
+
+        y_bar = y.mean('freq')
+        x_bar = x.mean('freq')
+
+        # using the formula to calculate the slope and intercept
+        n = np.sum((x - x_bar) * (y - y_bar), axis=0)
+        d = np.sum((x - x_bar)**2, axis=0)
+
+        m = n/d
+        b = y_bar - m*x_bar
+
+        return m, b
 
     def calc_epsilon_LT83(self, psd, U_mag, freq_range=[6.28, 12.57]):
         """Calculate the dissipation rate from the PSD
 
         Parameters
         ----------
-        psd : xarray.DataArray (...,time,f)
+        psd : xarray.DataArray (...,time,freq)
           The power spectral density
         U_mag : xarray.DataArray (...,time)
           The bin-averaged horizontal velocity [m s-1] (from dataset shortcut)
@@ -210,6 +339,7 @@ class ADVBinner(VelBinner):
         LT83 : Lumley and Terray, "Kinematics of turbulence convected
         by a random wave field". JPO, 1983, vol13, pp2000-2007.
         """
+
         # Ensure time has been averaged
         if len(psd.time)!=len(U_mag.time):
             raise Exception("`U_mag` should be from ensembled-averaged dataset")
@@ -228,12 +358,12 @@ class ADVBinner(VelBinner):
                freq.isel(freq=idx)**(5/3) / a).mean(axis=-1)**(3/2) / U
 
         return xr.DataArray(
-          out.astype('float32'),
-          attrs={'units': 'm2 s-3',
-                  'long_name': 'TKE Dissipation Rate',
-                  'standard_name': 'specific_turbulent_kinetic_energy_dissipation_in_sea_water',
-                  'description': 'TKE dissipation rate calculated using the method from Lumley and Terray, 1983',
-                  })
+            out.astype('float32'),
+            attrs={'units': 'm2 s-3',
+                   'long_name': 'Dissipation Rate',
+                   'standard_name': 'specific_turbulent_kinetic_energy_dissipation_in_sea_water',
+                   'description': 'TKE dissipation rate calculated using the method from Lumley and Terray, 1983',
+                   })
 
     def calc_epsilon_SF(self, vel_raw, U_mag, fs=None, freq_range=[2., 4.]):
         """Calculate dissipation rate using the "structure function" (SF) method
@@ -279,14 +409,14 @@ class ADVBinner(VelBinner):
             out[slc[:-1]] = (cv2m / 2.1) ** (3 / 2)
 
         return xr.DataArray(
-          out.astype('float32'),
-          coords=U_mag.coords,
-          dims=U_mag.dims,
-          attrs={'units': 'm2 s-3',
-                  'long_name': 'TKE Dissipation Rate',
-                  'standard_name': 'specific_turbulent_kinetic_energy_dissipation_in_sea_water',
-                  'description': 'TKE dissipation rate calculated using the "structure function" method',
-                  })
+            out.astype('float32'),
+            coords=U_mag.coords,
+            dims=U_mag.dims,
+            attrs={'units': 'm2 s-3',
+                   'long_name': 'Dissipation Rate',
+                   'standard_name': 'specific_turbulent_kinetic_energy_dissipation_in_sea_water',
+                   'description': 'TKE dissipation rate calculated using the "structure function" method',
+                   })
 
     def _up_angle(self, U_complex):
         """Calculate the angle of the turbulence fluctuations.
@@ -381,14 +511,14 @@ class ADVBinner(VelBinner):
         out *= 0.5
 
         return xr.DataArray(
-          out.astype('float32'),
-          coords={'time': dat_avg.psd.time},
-          dims='time',
-          attrs={'units': 'm2 s-3',
-                  'long_name': 'TKE Dissipation Rate',
-                  'standard_name': 'specific_turbulent_kinetic_energy_dissipation_in_sea_water',
-                  'description': 'TKE dissipation rate calculated using the method from Trowbridge and Elgar, 2001'
-                  })
+            out.astype('float32'),
+            coords={'time': dat_avg.psd.time},
+            dims='time',
+            attrs={'units': 'm2 s-3',
+                   'long_name': 'Dissipation Rate',
+                   'standard_name': 'specific_turbulent_kinetic_energy_dissipation_in_sea_water',
+                   'description': 'TKE dissipation rate calculated using the method from Trowbridge and Elgar, 2001'
+                   })
 
     def calc_L_int(self, a_cov, U_mag, fs=None):
         """Calculate integral length scales.
@@ -421,11 +551,12 @@ class ADVBinner(VelBinner):
         scale = np.argmin((acov/acov[..., :1]) > (1/np.e), axis=-1)
         L_int = U_mag.values / fs * scale
 
-        return xr.DataArray(L_int.astype('float32'),
-                            coords={'dir': a_cov.dir, 'time': a_cov.time},
-                            attrs={'units': 'm',
-                                   'long_name': 'Integral Length Scale',
-                                   'standard_name': 'turbulent_mixing_length_of_sea_water'})
+        return xr.DataArray(
+            L_int.astype('float32'),
+            coords={'dir': a_cov.dir, 'time': a_cov.time},
+            attrs={'units': 'm',
+                   'long_name': 'Integral Length Scale',
+                   'standard_name': 'turbulent_mixing_length_of_sea_water'})
 
 
 def calc_turbulence(ds_raw, n_bin, fs, n_fft=None, freq_units='rad/s', window='hann'):