+++ dolfyn/rotate/vector.py
@@ -62,7 +62,7 @@ def _inst2earth(advo, reverse=False, rotate_vars=None, force=False):
       If True, this function performs the inverse rotation (earth->inst).
     rotate_vars : iterable (default: None, list in advo.rotate_vars)
       The list of variables to rotate. By default this is taken from
-      advo.props['rotate_vars'].
+      advo.attrs['rotate_vars'].
     force : bool (default: False)
       Do not check which frame the data is in prior to performing 
       this rotation.
@@ -79,11 +79,7 @@ def _inst2earth(advo, reverse=False, rotate_vars=None, force=False):
         cs_now = 'inst'
         cs_new = 'earth'
 
-    if rotate_vars is None:
-        if 'rotate_vars' in advo.attrs:
-            rotate_vars = advo.rotate_vars
-        else:
-            rotate_vars = ['vel']
+    rotate_vars = rotb._check_rotate_vars(advo, rotate_vars)
 
     cs = advo.coord_sys.lower()
     if not force:
@@ -126,7 +122,7 @@ def _inst2earth(advo, reverse=False, rotate_vars=None, force=False):
     return advo
 
 
-def _earth2principal(advo, reverse=False):
+def _earth2principal(advo, reverse=False, rotate_vars=None):
     """Rotate data in an ADV dataset to/from principal axes. Principal
     heading must be within the dataset.
 
@@ -143,6 +139,9 @@ def _earth2principal(advo, reverse=False):
     reverse : bool (default: False)
       If True, this function performs the inverse rotation
       (principal->earth).
+    rotate_vars : iterable (default: None, list in advo.rotate_vars)
+      The list of variables to rotate. By default this is taken from
+      advo.props['rotate_vars'].
     """
 
     # This is in degrees CW from North
@@ -157,6 +156,8 @@ def _earth2principal(advo, reverse=False):
         ang *= -1
         cs_now = 'earth'
         cs_new = 'principal'
+
+    rotate_vars = rotb._check_rotate_vars(advo, rotate_vars)
 
     cs = advo.coord_sys.lower()
     if cs == cs_new:
@@ -174,7 +175,7 @@ def _earth2principal(advo, reverse=False):
                        [0, 0, 1]], dtype=np.float32)
 
     # Perform the rotation:
-    for nm in advo.rotate_vars:
+    for nm in rotate_vars:
         dat = advo[nm].values
         dat[:2] = np.einsum('ij,j...->i...', rotmat[:2, :2], dat[:2])
         advo[nm].values = dat.copy()