+++ dolfyn/tests/test_analysis.py
@@ -2,6 +2,7 @@ from dolfyn.tests import test_read_adp as tr, test_read_adv as tv
 from dolfyn.tests.base import load_netcdf as load, save_netcdf as save, assert_allclose
 from dolfyn import VelBinner, read_example
 import dolfyn.adv.api as avm
+import dolfyn.adp.api as apm
 from xarray.testing import assert_identical
 import pytest
 import numpy as np
@@ -24,41 +25,41 @@ class adp_setup():
 
 
 def test_do_func(make_data=False):
-    adv = adv_setup(tv)
-    bnr1 = adv.avg_tool
-    ds1 = adv.dat1
-    ds_vec = bnr1.do_avg(ds1)
-    ds_vec = bnr1.do_var(ds1, out_ds=ds_vec)
-    # test non-integer bin sizes
-    mean_test = bnr1.mean(ds1['vel'].values, n_bin=ds1.fs*1.01)
+    adv_tool = adv_setup(tv).avg_tool
+    ds1 = adv_setup(tv).dat1
+    ds_vec = adv_tool.do_avg(ds1)
+    ds_vec = adv_tool.do_var(ds1, out_ds=ds_vec)
 
-    sig = adp_setup(tr)
-    bnr2 = sig.avg_tool
-    ds2 = sig.dat
-    ds_sig = bnr2.do_avg(ds2)
-    ds_sig = bnr2.do_var(ds2, out_ds=ds_sig)
+    # test non-integer bin sizes
+    mean_test = adv_tool.mean(ds1['vel'].values, n_bin=ds1.fs*1.01)
+
+    adp_tool = adp_setup(tr).avg_tool
+    ds2 = adp_setup(tr).dat
+    ds_sig = adp_tool.do_avg(ds2)
+    ds_sig = adp_tool.do_var(ds2, out_ds=ds_sig)
 
     if make_data:
         save(ds_vec, 'vector_data01_avg.nc')
         save(ds_sig, 'BenchFile01_avg.nc')
         return
 
-    assert(np.sum(mean_test-ds_vec.vel.values) == 0, "Mean test failed")
+    assert np.sum(mean_test-ds_vec.vel.values) == 0, "Mean test failed"
     assert_allclose(ds_vec, load('vector_data01_avg.nc'), atol=1e-6)
     assert_allclose(ds_sig, load('BenchFile01_avg.nc'), atol=1e-6)
 
 
 def test_calc_func(make_data=False):
-    adv = adv_setup(tv)
-    bnr1 = adv.avg_tool
-    ds1 = adv.dat1
-    ds2 = adv.dat2
+    bnr1 = adv_setup(tv).avg_tool
+    ds1 = adv_setup(tv).dat1
+    ds2 = adv_setup(tv).dat2
 
-    sig = adp_setup(tr)
-    bnr = sig.avg_tool
-    ds = sig.dat
+    bnr = adp_setup(tr).avg_tool
+    ds = adp_setup(tr).dat
 
     ds_adv = type(ds1)()
+    ds_adv_dif = type(ds1)()
+    ds_adp = type(ds)()
+
     ds_adv['coh'] = bnr1.calc_coh(
         ds1['vel'][0], ds1['vel'][1], n_fft_coh=ds1.fs)
     ds_adv['pang'] = bnr1.calc_phase_angle(
@@ -70,14 +71,12 @@ def test_calc_func(make_data=False):
     ds_adv['psd'] = bnr1.calc_psd(ds1['vel'], freq_units='Hz')
 
     # Different lengths
-    ds_adv_dif = type(ds1)()
     ds_adv_dif['coh_dif'] = bnr1.calc_coh(
         ds1['vel'], ds2.vel)
     ds_adv_dif['pang_dif'] = bnr1.calc_phase_angle(
         ds1['vel'], ds2.vel)
 
     # Test ADCP single vector spectra, cross-spectra to test radians code
-    ds_adp = type(ds)()
     ds_adp['psd_b5'] = bnr.calc_psd(ds['vel_b5'].isel(
         range_b5=5), freq_units='rad', window='hamm')
     ds_adp['tke_b5'] = bnr.calc_tke(ds['vel_b5'])
@@ -105,23 +104,64 @@ def test_calc_freq():
 
 
 def test_adv_turbulence(make_data=False):
+    # Checks that code is functioning
     dat = tv.dat.copy(deep=True)
     bnr = avm.ADVBinner(n_bin=20.0, fs=dat.fs)
     tdat = bnr(dat)
-    acov = bnr.calc_acov(dat.vel)
+    acov = bnr.calc_acov(dat['vel'])
 
     assert_identical(tdat, avm.calc_turbulence(dat, n_bin=20.0, fs=dat.fs))
 
-    tdat['stress_detrend'] = bnr.calc_stress(dat.vel)
-    tdat['stress_demean'] = bnr.calc_stress(dat.vel, detrend=False)
-    tdat['csd'] = bnr.calc_csd(dat.vel, freq_units='rad', window='hamm')
-    tdat['LT83'] = bnr.calc_epsilon_LT83(tdat.psd, tdat.velds.U_mag)
-    tdat['SF'] = bnr.calc_epsilon_SF(dat.vel[0], tdat.velds.U_mag)
+    tdat['stress_detrend'] = bnr.calc_stress(dat['vel'])
+    tdat['stress_demean'] = bnr.calc_stress(dat['vel'], detrend=False)
+    tdat['csd'] = bnr.calc_csd(
+        dat['vel'], freq_units='rad', window='hamm', n_fft_coh=10)
+    tdat['LT83'] = bnr.calc_epsilon_LT83(tdat['psd'], tdat.velds.U_mag)
+    tdat['SF'] = bnr.calc_epsilon_SF(dat['vel'][0], tdat.velds.U_mag)
     tdat['TE01'] = bnr.calc_epsilon_TE01(dat, tdat)
     tdat['L'] = bnr.calc_L_int(acov, tdat.velds.U_mag)
+    slope_check = bnr.check_turbulence_cascade_slope(
+        tdat['psd'][-1].mean('time'), freq_range=[10, 100])
 
     if make_data:
         save(tdat, 'vector_data01_bin.nc')
         return
 
+    assert np.round(slope_check[0].values, 4), 0.1713
     assert_allclose(tdat, load('vector_data01_bin.nc'), atol=1e-6)
+
+
+def test_adcp_turbulence(make_data=False):
+    dat = tr.dat_sig_i.copy(deep=True)
+    bnr = apm.ADPBinner(n_bin=20.0, fs=dat.fs, diff_style='centered')
+    tdat = bnr.do_avg(dat)
+    tdat['dudz'] = bnr.calc_dudz(tdat.vel)
+    tdat['dvdz'] = bnr.calc_dvdz(tdat.vel)
+    tdat['dwdz'] = bnr.calc_dwdz(tdat.vel)
+    tdat['tau2'] = bnr.calc_shear2(tdat.vel)
+    tdat['psd'] = bnr.calc_psd(dat['vel'].isel(
+        dir=2, range=len(dat.range)//2), freq_units='Hz')
+    tdat['noise'] = bnr.calc_doppler_noise(tdat['psd'], pct_fN=0.8)
+    tdat['stress_vec4'] = bnr.calc_stress_4beam(
+        dat, noise=tdat['noise'], orientation='up', beam_angle=25)
+    tdat['tke_vec5'], tdat['stress_vec5'] = bnr.calc_stress_5beam(
+        dat, noise=tdat['noise'], orientation='up', beam_angle=25, tke_only=False)
+    tdat['tke'] = bnr.calc_total_tke(
+        dat, noise=tdat['noise'], orientation='up', beam_angle=25)
+    # This is "negative" for this code check
+    tdat['wpwp'] = bnr.calc_tke(dat['vel_b5'], noise=tdat['noise'])
+    tdat['dissipation_rate_LT83'] = bnr.calc_dissipation_LT83(
+        tdat['psd'], tdat.velds.U_mag.isel(range=len(dat.range)//2), freq_range=[0.2, 0.4])
+    tdat['dissipation_rate_SF'], tdat['noise_SF'], tdat['D_SF'] = bnr.calc_dissipation_SF(
+        dat.vel.isel(dir=2), r_range=[1, 5])
+    tdat['friction_vel'] = bnr.calc_ustar_fit(
+        tdat, upwp_=tdat['stress_vec5'].sel(tau='upwp_'), z_inds=slice(1, 5), H=50)
+    slope_check = bnr.check_turbulence_cascade_slope(
+        tdat['psd'].mean('time'), freq_range=[0.4, 4])
+
+    if make_data:
+        save(tdat, 'Sig1000_IMU_bin.nc')
+        return
+
+    assert np.round(slope_check[0].values, 4), -1.0682
+    assert_allclose(tdat, load('Sig1000_IMU_bin.nc'), atol=1e-6)