+++ dolfyn/tests/test_clean.py
@@ -3,7 +3,6 @@ from dolfyn.tests import test_read_adp as tp
 from dolfyn.tests.base import load_netcdf as load, save_netcdf as save, assert_allclose
 import dolfyn.adv.api as avm
 import dolfyn.adp.api as apm
-import numpy as np
 
 
 def test_GN2002(make_data=False):
@@ -61,7 +60,7 @@ def test_clean_upADCP(make_data=False):
     td_sig = tp.dat_sig_tide.copy(deep=True)
 
     apm.clean.find_surface_from_P(td_awac, salinity=30)
-    td_awac = apm.clean.nan_beyond_surface(td_awac)
+    td_awac = apm.clean.nan_beyond_surface(td_awac, beam_angle=20)
 
     apm.clean.set_range_offset(td_sig, 0.6)
     apm.clean.find_surface_from_P(td_sig, salinity=31)